requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
Pillow==10.3.0
